Rename file to itk_torch_bridge
