+++ monai/data/itk_torch_bridge.py
